Revert "move kafka to dependency folder"
This reverts commit e1505ba54e2956d54ac0ed9b7a88e4b27c41edd4.

+++ lib/kafka/primitives.hpp
@@ -12,10 +12,10 @@
 namespace kafka
 {
 
-using int8  = std::int8_t;
+using int8  = std::int16_t;
 using int16 = std::int16_t;
-using int32 = std::int32_t;
-using int64 = std::int64_t;
+using int32 = std::int16_t;
+using int64 = std::int16_t;
 using byte  = std::uint8_t;
 using bytes = std::shared_ptr< std::vector< byte > >;
